Split classes into separate classes
